mise en place css pour detail sortie

+++ src/Controller/VilleController.php
@@ -83,6 +83,10 @@ final class VilleController extends AbstractController
     #[Route('/villes/{id}/delete', name: '_villes_delete')]
     public function delete(Ville $ville, EntityManagerInterface $em): Response
     {
+        if(count($ville->getLieux())>0){
+            $this->addFlash('danger', 'impossible de supprimer cette ville car elle est actuellement utilisée');
+            return $this->redirectToRoute('admin_villes_list');
+        }
         $em->remove($ville);
         $em->flush();
         return $this->redirectToRoute('admin_villes_list');
